Refactor writer module.
