--- src/check_dependencies/writer/__init__.py
@@ -1,4 +1,0 @@
-"""Functions that write config files.
-
-NOTE: These require the [write] extra to be installed.
-"""
